+++ deltafi-ui/src/store/modules/uiConfig/state.ts
@@ -3,6 +3,12 @@ export type State = {
   uiConfig: {
     title: String,
     domain: String,
+    securityBanner: {
+      enabled: Boolean,
+      backgroundColor: String,
+      textColor: String
+      text: String
+    }
     dashboard: {
       links: Array<{
         name: String,
@@ -18,6 +24,12 @@ export const state: State = {
   uiConfig: {
     title: 'DeltaFi',
     domain: 'example.deltafi.org',
+    securityBanner: {
+      enabled: false,
+      backgroundColor: "#FFFFFFF",
+      textColor: "#000000",
+      text: 'DeltaFi',
+    },
     dashboard: {
       links: []
     }